+++ src/collections/contents/server/methods.js
@@ -1,12 +1,22 @@
+/*
+ * pwix:editor/src/collections/contents/server/methods.js
+ */
+
+import _ from 'lodash';
+const assert = require( 'assert' ).strict; // up to nodejs v16.x
+
 import { Contents } from '../contents.js';
 
 Meteor.methods({
     // empty the collection
-    'te_contents.empty'(){
+    'te_contents.empty'( collection_name ){
+        assert( collection_name && _.isString( collection_name ) && collection_name.length, 'te_contents.empty expects collection_name be a non-empty string' );
+
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
-            console.debug( 'pwix:editor te_contents.empty() call' );
+            console.debug( 'pwix:editor te_contents.empty() call with collection='+collection_name );
         }
-        const res = Editor.collections.Contents.server.remove({});
+        const collection = Editor.collections.get( collection_name, Editor.collections.Contents.schema );
+        const res = collection.remove({});
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.empty() returns with', res );
         }
@@ -14,11 +24,15 @@ Meteor.methods({
     },
 
     // get some content from the database
-    'te_contents.byName'( name ){
+    'te_contents.byName'( collection_name, document_name ){
+        assert( collection_name && _.isString( collection_name ) && collection_name.length, 'te_contents.byName expects collection_name be a non-empty string' );
+        assert( document_name && _.isString( document_name ) && document_name.length, 'te_contents.byName expects document_name be a non-empty string' );
+
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
-            console.debug( 'pwix:editor te_contents.byName() call with name=', name );
+            console.debug( 'pwix:editor te_contents.byName() call with collection='+collection_name, 'document='+document_name );
         }
-        const res = Editor.collections.Contents.server.findOne({ name: name });
+        const collection = Editor.collections.get( collection_name, Editor.collections.Contents.schema );
+        const res = collection.findOne({ name: document_name });
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.byName() returns with', res );
         }
@@ -26,11 +40,14 @@ Meteor.methods({
     },
 
     // import an element (so do not modify any data) 
-    'te_contents.import'( elt ){
+    'te_contents.import'( collection_name, elt ){
+        assert( collection_name && _.isString( collection_name ) && collection_name.length, 'te_contents.import expects collection_name be a non-empty string' );
+
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.import() call with elt=', elt );
         }
-        const res = Editor.collections.Contents.server.insert( elt );
+        const collection = Editor.collections.get( collection_name, Editor.collections.Contents.schema );
+        const res = collection.insert( elt );
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.import() returns with', res );
         }
@@ -39,16 +56,15 @@ Meteor.methods({
 
     // set some content into the database
     // name is mandatory
-    'te_contents.set'( name, content ){
+    'te_contents.set'( collection_name, document_name, content ){
+        assert( collection_name && _.isString( collection_name ) && collection_name.length, 'te_contents.byName expects collection_name be a non-empty string' );
+        assert( document_name && _.isString( document_name ) && document_name.length, 'te_contents.byName expects document_name be a non-empty string' );
+
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
-            console.debug( 'pwix:editor te_contents.set() call with name=', name, 'content=', content );
+            console.debug( 'pwix:editor te_contents.set() call with collection='+collection_name, 'document='+document_name, 'content=', content );
         }
-        if( !name ){
-            throw new Meteor.Error(
-                'content.set',
-                'Empty name' );
-        }
-        let orig = Editor.collections.Contents.server.findOne({ name: name });
+        const collection = Editor.collections.get( collection_name, Editor.collections.Contents.schema );
+        let orig = collection.findOne({ name: document_name });
         //console.log( 'content.set orig=', orig );
         let o = {
             content: content
@@ -62,9 +78,9 @@ Meteor.methods({
             o.createdAt = new Date();
             o.createdBy = this.userId;
         }
-        let res = Editor.collections.Contents.server.upsert({ name: name }, { $set: o });
+        let res = collection.upsert({ name: document_name }, { $set: o });
         if( res.numberAffected > 0 ){
-            res.written = Editor.collections.Contents.server.findOne({ name: name });
+            res.written = collection.findOne({ name: document_name });
         }
         if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.set() returns with', res );