+++ src/server/js/startup.js
@@ -1,22 +1,3 @@
 /*
  * pwix:editor/src/server/js/startup.js
  */
-
-import { Mongo } from 'meteor/mongo';
-
-Meteor.startup( function(){
-    // define the server-side collections
-    if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
-        console.debug( 'pwix:editor defining server-side collections...' );
-    }
-    Object.keys( Editor.collections ).every(( c ) => {
-        const _name = Editor._conf.collections.prefix + Editor.collections[c].name;
-        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
-            console.debug( '   '+c+' -> '+_name );
-        }
-        Editor.collections[c].server = new Mongo.Collection( _name );
-        Editor.collections[c].server.attachSchema( Editor.collections[c].schema );
-        Editor.collections[c].deny();
-        return true;
-    });
-});